requirements.txt
flask
mysql-connector-python
flask-cors
werkzeug


